requirements.txt
pydantic>=2.0
pydantic-settings
python-dotenv
chromadb
openai
deepgram-sdk
feedparser
yt-dlp
httpx
boto3
feedgen